+++ plugins/commentable-html/dev/assets/js/83-print.js
@@ -294,6 +294,7 @@ function setupSinglePagePrint() {
   function measureCss() {
     return ".cmh-print-comments,.cmh-print-noscript{display:block !important}"
       + "#commentRoot section.cmh-section-collapsed>*{display:revert !important}"
+      + "#commentRoot section.cm-toc-filtered{display:revert !important}"
       + "#commentRoot .cmh-note.cmh-note-collapsed .cmh-note-input,"
       + "#commentRoot .cmh-note.cmh-note-collapsed .cmh-note-head{display:revert !important}"
       + "#commentRoot pre,#commentRoot code,#commentRoot .cmh-diff-view pre,#commentRoot .cmh-diff-view code,"